% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RACT.R
\name{calc_ky_fan_k}
\alias{calc_ky_fan_k}
\title{Calculate Ky-Fan(k) norm}
\usage{
calc_ky_fan_k(cov_cor_X_1, cov_cor_X_2, K)
}
\arguments{
\item{cov_cor_X_1}{p x p (p dimensions) covariance or correlation matrix for group 1}

\item{cov_cor_X_2}{p x p  (p dimensions) covariance or correlation matrix for group 2}

\item{K}{Vector of integers, and these integers represent the Ky-Fan(k) norms that are included.
E.g. K=c(1,5) means RACT is based on the Ky-Fan(1) and Ky-Fan(5) norm.}
}
\value{
named vector of Ky-Fan(k) norms where names of each entry are values of k
}
\description{
This function takes in two covariance/correlation matrices, and calculates the Ky-Fan(k) norm of their differences for k=1,...,K
}
