% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RACT.R
\name{K_calculate}
\alias{K_calculate}
\title{This function calculates the smallest K, such that the sum of the top K singular values of the covariance/correlation
  matrix of the data is at least K_pct of the sum of all singular values.}
\usage{
K_calculate(X_1, X_2, K_pct = 0.8, cov = TRUE)
}
\arguments{
\item{X_1}{p x n data matrix (p dimensions, n observations) for group 1}

\item{X_2}{p x n data matrix (p dimensions, n observations) for group 2}

\item{K_pct}{0 < K_pct <= 1, where K_pct determines how large K should be so that sum of top K singular values of pooled covariance/correlation at least
K_pct of the sum of all singular values. Default set at 0.8.}

\item{cov}{If cov=TRUE then K based off of pooled covariance, if cov=FALSE then based off of pooled correlation.}
}
\value{
vector of values c(1,...,K)
}
\description{
This function takes in two covariance/correlation matrices, and calculates the Ky-Fan(k) norm of their differences for k=1,...,K
}
