% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RACT.R
\name{RACT}
\alias{RACT}
\title{RACT}
\usage{
RACT(X_1, X_2, n_perm = 1000, K = NULL, min_P = FALSE, cov = TRUE, seed = NULL)
}
\arguments{
\item{X_1}{p x n data matrix (p dimensions, n observations) for group 1}

\item{X_2}{p x n data matrix (p dimensions, n observations) for group 2}

\item{n_perm}{number of permutations to use in the permutation procedure}

\item{K}{If NULL, Ky-Fan(k) norms from 1 to K are included where K smallest integer such that sum of top K singular values of the pooled covariance
are greater than 80% of the sum of all of the singular values. Otherwise this argument takes a vector of integers, and these integers represent the
Ky-Fan(k) norms that are included. E.g. K=c(1,5) means RACT is based on the Ky-Fan(1) and Ky-Fan(5) norms only.}

\item{min_P}{Whether a minimum-P value statistic is used.}

\item{cov}{If cov=TRUE RACT tests for the equality of the covariance of X_1,X_2. If cov=FALSE RACT tests for the equality of the correlation of X_1,X_2.}

\item{seed}{Allows user to set a seed for the permutation procedure.}
}
\value{
a list of length 2. 'adaptive_p_value' is the p-value associated with RACT's adaptive test statistic. 'p_value_vector' a vector with p-values
  associated with individual Ky-Fan(k) norms
}
\description{
This function takes in two data matrices, and tests for the equality of covariance using rank-adaptive covariance testing.
}
\examples{
X_1 =
}
